Revert back refactoring into separate step
